+++ src/commands/index.ts
@@ -3,6 +3,7 @@ import Reminder from './reminder';
 import Help from './help';
 import BaseCommand from './base';
 import About from './about';
+import Uptime from './uptime';
 
 export interface CommandMap {
     [key: string]: BaseCommand;
@@ -13,4 +14,5 @@ export default {
     reminder: new Reminder(),
     help: new Help(),
     about: new About(),
+    uptime: new Uptime(),
 } as CommandMap;