+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added: Reminder system via the `!reminders` command.
 * Added: If a subcommand is missing, help will be displayed for that particular command.
 * Added: The `!help` command now shows commands with available help.
+* Added: `!about` and `!uptime` commands.
 
 ## [v0.1.0]